shorten path to solve deployement problem
