--- js/UserInterface/UiElements/BlockerScreen/Menu/MenuElements/MenuDisplay/MenuPanels/MultiplayerPanel/MultiplayerPanelElements/GoLiveDisplay/GoLiveDisplayElements/GoLiveButton.js
@@ -1,13 +1,0 @@
-import { UiElement } from "../../../../../../../../../UiElement.js";
-
-class GoLiveButton extends UiElement {
-    constructor(onClick) {
-        super({
-            id: "GoLiveButton",
-            type: "button",
-            innerHTML: "Go live",
-            onClick: onClick
-        })
-    }
-}
-export { GoLiveButton }
