--- js/UserInterface/UiElements/BlockerScreen/Menu/MenuElements/MenuDisplay/MenuPanels/MultiplayerPanel/MultiplayerPanelElements/GoLiveDisplay/GoLiveDisplayElements/PeerIdInput.js
@@ -1,11 +1,0 @@
-import { UiElement } from "../../../../../../../../../UiElement.js";
-
-class PeerIdInput extends UiElement { 
-    constructor() {
-        super({
-            id: "PeerIdInput",
-            type: "input"
-        })
-    }
-}
-export { PeerIdInput }
