+++ js/UserInterface/UiElements/MenuPanels/ChatPanel/ChatPanelElements/ChatOutputDisplay.js
@@ -1,0 +1,22 @@
+import { UiElement } from "../../../UiElement.js";
+
+class ChatOutputDisplay extends UiElement {
+    constructor(){
+        super({
+            id: "ChatOutputDisplay",
+            style: {
+                height: "600px",
+                background: "white",
+                overflow: 'scroll',
+                overflowX: 'hidden',
+                wordBreak: 'break-all',
+                fontSize: "x-small",
+            }
+        })
+    }
+
+    displayNewMessage(message) {
+        this.element.innerHTML += "<br>"+message;
+    }
+}
+export { ChatOutputDisplay }
