+++ js/UserInterface/UiElements/MenuPanels/MapPanel/MapPanelElements/graphqlCaller.js
