+++ js/UserInterface/UiElements/MenuPanels/MultiplayerPanel/MultiplayerPanelElements/GoLiveDisplay/GoLiveDisplayElements/PeerIdInput.js
@@ -1,0 +1,11 @@
+import { UiElement } from "../../../../../UiElement.js";
+
+class PeerIdInput extends UiElement { 
+    constructor() {
+        super({
+            id: "PeerIdInput",
+            type: "input"
+        })
+    }
+}
+export { PeerIdInput }
